Add starting point to indexed

+++ src/main/java/com/ginsberg/gatherers4j/IndexedValue.java
@@ -16,5 +16,5 @@
 
 package com.ginsberg.gatherers4j;
 
-public record IndexedValue<TYPE>(int index, TYPE value) {
+public record IndexedValue<TYPE>(long index, TYPE value) {
 }
